requirements.txt
anyio==3.6.2
appnope==0.1.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
black==23.3.0
bleach==6.0.0
bokeh==2.4.3
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comm==0.1.3
concurrent-log-handler==0.9.24
cx_Oracle==8.3.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Django==5.0
django-audit-fields==0.3.9
django-defender==0.9.7
django-environ==0.10.0
django-revision==0.3.6
django-simple-history==3.3.0
django-tastypie==0.14.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
edc-device==0.3.10
edc-utils==0.3.19
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.16.3
flake8==6.0.0
fqdn==1.5.1
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
Markdown==3.4.3
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
mssql==1.0.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
portalocker==2.7.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyct==0.5.0
pyflakes==3.0.1
Pygments==2.15.1
PyJWT==2.8.0
pyodbc==4.0.39
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-mimeparse==1.6.0
pytz==2023.3
pyviz-comms==2.2.1
PyYAML==6.0.1
pyzmq==25.0.2
qtconsole==5.4.3
QtPy==2.3.1
redis==4.6.0
requests==2.30.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.12
sqlparse @ file:///private/var/folders/k1/30mswbxs7r1g6zwn8y4fyt500000gp/T/abs_16ptvrhamb/croot/sqlparse_1690901943673/work
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tomli==2.0.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==2.0.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
widgetsnbextension==4.0.7
xlrd==2.0.1
XlsxWriter==3.1.2
zipp==3.15.0
